requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
environs==9.5.0
frozenlist==1.3.3
h3==3.7.6
idna==3.4
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
mysql-connector-python==8.0.32
numpy==1.24.3
packaging==23.0
peewee==3.16.0
prettytable==3.7.0
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
timezonefinder==6.2.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.8.2
